Change the structure so that the challenge integrates the auto-logging.
