+++ .settings/org.python.pydev.analysis.yaml
@@ -1,0 +1,5 @@
+MYPY_USE_CONSOLE: false
+SEARCH_MYPY_LOCATION: SEARCH
+USE_MYPY: true
+MYPY_ADD_PROJECT_FOLDERS_TO_MYPYPATH: true
+MYPY_FILE_LOCATION: ''
