+++ .settings/org.python.pydev.yaml
@@ -1,0 +1,38 @@
+ADD_COMMENTS_AT_INDENT: false
+ADD_NEW_LINE_AT_END_OF_FILE: true
+AUTOPEP8_PARAMETERS: ''
+BLACK_PARAMETERS: --fast
+BLANK_LINES_INNER: 1
+BLANK_LINES_TOP_LEVEL: 2
+DATE_FIELD_FORMAT: yyyy-MM-dd
+DATE_FIELD_NAME: __updated__
+ENABLE_DATE_FIELD_ACTION: false
+FORMATTER_STYLE: BLACK
+FORMAT_BEFORE_SAVING: true
+FORMAT_ONLY_CHANGED_LINES: false
+MANAGE_BLANK_LINES: true
+MULTI_BLOCK_COMMENT_CHAR: =
+MULTI_BLOCK_COMMENT_SHOW_ONLY_CLASS_NAME: true
+MULTI_BLOCK_COMMENT_SHOW_ONLY_FUNCTION_NAME: true
+MYPY_ADD_PROJECT_FOLDERS_TO_MYPYPATH: true
+MYPY_ARGS: --follow-imports=silent --show-column-numbers
+MYPY_FILE_LOCATION: ''
+MYPY_USE_CONSOLE: true
+PYDEV_TEST_RUNNER: '2'
+PYDEV_TEST_RUNNER_DEFAULT_PARAMETERS: --capture=no -W ignore::DeprecationWarning -n
+  0 --tb=native -vv --force-regen
+PYDEV_USE_PYUNIT_VIEW: true
+SAVE_ACTIONS_ONLY_ON_WORKSPACE_FILES: true
+SEARCH_MYPY_LOCATION: SEARCH
+SINGLE_BLOCK_COMMENT_ALIGN_RIGHT: true
+SINGLE_BLOCK_COMMENT_CHAR: '-'
+SORT_IMPORTS_ON_SAVE: false
+SPACES_BEFORE_COMMENT: '2'
+SPACES_IN_START_COMMENT: '1'
+TRIM_EMPTY_LINES: true
+TRIM_MULTILINE_LITERALS: true
+USE_ASSIGN_WITH_PACES_INSIDER_PARENTESIS: false
+USE_MYPY: true
+USE_OPERATORS_WITH_SPACE: true
+USE_SPACE_AFTER_COMMA: true
+USE_SPACE_FOR_PARENTESIS: false
