--- examples/rpa-challenge/pyproject.toml
@@ -1,12 +1,0 @@
-[tool.robo]
-name = "Simple Automation"
-description = "Very simple automation "
-readme = "README.md"
-
-[tool.robo.dependencies]
-rpaframework="22.0.0"
-
-# Do these come here or do they come from robo
-[tool.robo.dev-dependencies]
-ruff=""
-black=""
